define config in cmd folder
