--- internal/config/config.go
@@ -1,56 +1,0 @@
-package config
-
-type Config struct {
-	RpcConfig      RPCConfig
-	RedisConfig    RedisConfig
-	HorizonConfig  HorizonConfig
-	PricesConfig   PricesConfig
-	BlockaidConfig BlockaidConfig
-	CoinbaseConfig CoinbaseConfig
-	AppConfig      AppConfig
-}
-
-type AppConfig struct {
-	Mode      string
-	SentryKey string
-}
-
-type RPCConfig struct {
-	RpcPubnetURL  string
-	RpcTestnetURL string
-}
-
-type RedisConfig struct {
-	ConnectionName string
-	Host           string
-	Port           int
-}
-
-type HorizonConfig struct {
-	HorizonPubnetURL  string
-	HorizonTestnetURL string
-}
-
-type PricesConfig struct {
-	HorizonURL                     string
-	DisableTokenPrices             bool
-	BatchUpdateDelayMilliseconds   int
-	CalculationTimeoutMilliseconds int
-	UpdateIntervalMilliseconds     int
-	UpdateBatchSize                int
-	StalenessThreshold             int
-}
-
-type BlockaidConfig struct {
-	BlockaidAPIKey                         string
-	UseBlockaidDappScanning                bool
-	UseBlockaidTxScanning                  bool
-	UseBlockaidAssetScanning               bool
-	UseBlockaidAssetWarningReporting       bool
-	UseBlockaidTransactionWarningReporting bool
-}
-
-type CoinbaseConfig struct {
-	CoinbaseAPIKey    string
-	CoinbaseAPISecret string
-}
